backend/requirements.txt
flask
flask-jwt-extended
flask-cors
pymongo
python-dotenv
werkzeug
marshmallow
flask-marshmallow
email-validator
dnspython
